# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  dto.CreateAPIKeyRequest:
    properties:
      expires_at:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      type:
        allOf:
        - $ref: '#/definitions/types.SecretType'
        enum:
        - private_key
        - publishable_key
    required:
    - name
    - type
    type: object
  dto.CreateAPIKeyResponse:
    properties:
      api_key:
        type: string
      secret:
        $ref: '#/definitions/dto.SecretResponse'
    type: object
...
  dto.CreateIntegrationRequest:
    properties:
      credentials:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      provider:
        $ref: '#/definitions/types.SecretProvider'
    required:
    - credentials
    - name
    - provider
    type: object
...
  dto.LinkedIntegrationsResponse:
    properties:
      providers:
        items:
          type: string
        type: array
    type: object
...
  dto.ListSecretsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.SecretResponse'
        type: array
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
    type: object
...
  dto.SecretResponse:
    properties:
      created_at:
        type: string
      display_id:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      provider:
        $ref: '#/definitions/types.SecretProvider'
      status:
        $ref: '#/definitions/types.Status'
      type:
        $ref: '#/definitions/types.SecretType'
      updated_at:
        type: string
    type: object
...
  errors.ErrorDetail:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorDetail'
      success:
        type: boolean
    type: object
...
  types.PaginationResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  types.SecretProvider:
    enum:
    - flexprice
    - stripe
    - razorpay
    type: string
    x-enum-varnames:
    - SecretProviderFlexPrice
    - SecretProviderStripe
    - SecretProviderRazorpay
  types.SecretType:
    enum:
    - private_key
    - publishable_key
    - integration
    type: string
    x-enum-varnames:
    - SecretTypePrivateKey
    - SecretTypePublishableKey
    - SecretTypeIntegration
  types.Status:
    enum:
    - published
    - deleted
    - archived
    type: string
    x-enum-varnames:
    - StatusPublished
    - StatusDeleted
    - StatusArchived
...
info:
  contact: {}
  description: FlexPrice API Service
  title: FlexPrice API
  version: "1.0"
paths:
...
  /secrets/api/keys:
    get:
      consumes:
      - application/json
      description: Get a paginated list of API keys
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Status (published/archived)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListSecretsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List API keys
      tags:
      - secrets
    post:
      consumes:
      - application/json
      description: Create a new API key with the specified type and permissions
      parameters:
      - description: API key creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new API key
      tags:
      - secrets
  /secrets/api/keys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an API key by ID
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Delete an API key
      tags:
      - secrets
  /secrets/integrations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete integration credentials
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Delete an integration
      tags:
      - secrets
  /secrets/integrations/{provider}:
    get:
      consumes:
      - application/json
      description: Get details of a specific integration
      parameters:
      - description: Integration provider
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SecretResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get integration details
      tags:
      - secrets
    post:
      consumes:
      - application/json
      description: Create or update integration credentials
      parameters:
      - description: Integration provider
        in: path
        name: provider
        required: true
        type: string
      - description: Integration creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateIntegrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SecretResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create or update an integration
      tags:
      - secrets
  /secrets/integrations/linked:
    get:
      consumes:
      - application/json
      description: Get a list of unique providers which have a valid linked integration
        secret
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LinkedIntegrationsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List linked integrations
      tags:
      - secrets
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key in the format *x-api-key &lt;api-key&gt;**
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"